neonwar beta 1.2:     performance upgrade

+++ jsBattle/classes.js
@@ -1,5 +1,5 @@
 function playerRender(obToControll) {
-  if (obToControll.xPos + playerSize - cameraX > 0 && obToControll.yPos + playerSize - cameraY > 0 && obToControll.xPos - playerSize - cameraX < xScreenSize && obToControll.yPos - playerSize - cameraY < yScreenSize) {
+  if (onScreen(obToControll)) {
     push();
       translate(cameraX + (obToControll.xPos - cameraX), cameraY + (obToControll.yPos - cameraY)); // move (0,0) to me
       rotate(-obToControll.fireDirection + PI); // rotate me
@@ -20,39 +20,46 @@ class AIPlayer {
     this.health = playerHealth;
     this.turnDirection = HALF_PI;
     this.fireDirection = 0;
+    this.nowTargeting = 0;
+    this.closestPlayerSQDist = Infinity;
     this.playerLook = floor(random(0,playerTextures.length));
+    this.frameOffset = floor(random(60));
   }
   tick() {
-    var closestPlayerID = undefined; // find closest enemy
-    var closestPlayerSQDist = Infinity;
-    for (var i = 0; i < players.length; i ++) {
-      if (players[i] != this) {
-        var nowSQDist = SQdist(this.xPos, this.yPos, players[i].xPos, players[i].yPos);
-        if (nowSQDist < closestPlayerSQDist) {
-          closestPlayerSQDist = nowSQDist;
-          closestPlayerID = i;
+    if ((frameCount+this.frameOffset)%30 == 0) {
+      this.closestPlayerSQDist = Infinity;
+      for (var i = 0; i < players.length; i ++) {
+        if (players[i] != this) {
+          var nowSQDist = SQdist(this.xPos, this.yPos, players[i].xPos, players[i].yPos);
+          if (nowSQDist < this.closestPlayerSQDist) {
+            this.closestPlayerSQDist = nowSQDist;
+            this.nowTargeting = i;
+          }
         }
       }
-    }
-    if (closestPlayerID !== undefined) { // if enemy exists
-      var direction = atan2(players[closestPlayerID].xPos - this.xPos, players[closestPlayerID].yPos - this.yPos); // aim at it
-      this.fireDirection = direction // remember its direction
-      if (closestPlayerSQDist < sq(500)) { // if enemy is close
-        direction += this.turnDirection; // move to the left/right instead
-      }
-      this.xSpeed += sin(direction)*walkSpeed; // move (accelerate)
-      this.ySpeed += cos(direction)*walkSpeed;
-
-      if (this.reload <= 0) { // if can fire
-        fireBullet(this.xPos, this.yPos, this.fireDirection, this.hue); // fire bullet
-        this.reload = reloadTime; // start reloading
+      if (floor(random(10)) === 0) { // swap between moving left/right if randomness feels like doing so
+        this.turnDirection = -this.turnDirection;
       }
     }
+    var direction = atan2(players[this.nowTargeting].xPos - this.xPos, players[this.nowTargeting].yPos - this.yPos); // aim at it
+    this.fireDirection = direction // remember its direction
+    if (this.closestPlayerSQDist < sq(500)) { // if enemy is close
+      direction += this.turnDirection; // move to the left/right instead
+    }
+    this.xSpeed += sin(direction)*walkSpeed; // move (accelerate)
+    this.ySpeed += cos(direction)*walkSpeed;
 
-    for (var i = 0; i < bullets.length; i ++) { // loop trought bullets
-      if (bullets[i].hue !== this.hue && SQdist(this.xPos, this.yPos, bullets[i].xPos, bullets[i].yPos) < sq(playerSize/2)) { // if bullet is not mine
-        this.health -= 10; // take damage
-        bullets[i].timeLeft = 0; // kill bullet
+    if (this.reload <= 0) { // if can fire
+      fireBullet(this.xPos, this.yPos, this.fireDirection, this.hue); // fire bullet
+      this.reload = reloadTime; // start reloading
+    }
+
+    if ((frameCount+this.frameOffset)%3 == 0) {
+      for (var i = 0; i < bullets.length; i ++) { // loop trought bullets
+        if (bullets[i].hue !== this.hue && SQdist(this.xPos, this.yPos, bullets[i].xPos, bullets[i].yPos) < sq(playerSize/2)) { // if bullet is not mine
+          this.health -= 10; // take damage
+          bullets[i].timeLeft = 0; // kill bullet
+        }
       }
     }
 
@@ -64,9 +71,6 @@ class AIPlayer {
     // if (this.health < 100) { // regenerate
     //   this.health += 0.1;
     // }
-    if (floor(random(500)) === 0) { // swap between moving left/right if randomness feels like doing so
-      this.turnDirection = -this.turnDirection;
-    }
   }
   render(){
     playerRender(this);
@@ -139,7 +143,7 @@ class bullet{
     this.timeLeft -= 1; // age
   }
   render() {
-    if (this.xPos + bulletSize - cameraX > 0 && this.yPos + bulletSize - cameraY > 0 && this.xPos - bulletSize - cameraX < xScreenSize && this.yPos - bulletSize - cameraY < yScreenSize) {
+    if (onScreen(this)) {
       fill(this.hue,100,50,this.timeLeft); // render
       rect(this.xPos, this.yPos, bulletSize, bulletSize);
     }
@@ -159,7 +163,7 @@ class particle{
     this.xPos += this.xSpeed; // move
     this.yPos += this.ySpeed;
     this.timeLeft -= 1; // age
-    if (this.xPos + bulletSize - cameraX > 0 && this.yPos + bulletSize - cameraY > 0 && this.xPos - bulletSize - cameraX < xScreenSize && this.yPos - bulletSize - cameraY < yScreenSize) {
+    if (onScreen(this)) {
       fill(this.hue, this.timeLeft, 50, this.timeLeft); // render
       rect(this.xPos, this.yPos, bulletSize, bulletSize);
     }
